+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiGridFilter.java
@@ -13,10 +13,12 @@ import java.io.IOException;
 public class GuiGridFilter extends GuiBase {
     private int compare;
     private int mode;
+    private boolean modFilter;
 
     private GuiCheckBox compareDamage;
     private GuiCheckBox compareNBT;
     private GuiCheckBox compareOredict;
+    private GuiCheckBox toggleModFilter;
     private GuiButton toggleMode;
 
     public GuiGridFilter(ContainerGridFilter container) {
@@ -24,6 +26,7 @@ public class GuiGridFilter extends GuiBase {
 
         this.compare = ItemGridFilter.getCompare(container.getStack());
         this.mode = ItemGridFilter.getMode(container.getStack());
+        this.modFilter = ItemGridFilter.isModFilter(container.getStack());
     }
 
     @Override
@@ -31,14 +34,16 @@ public class GuiGridFilter extends GuiBase {
         compareDamage = addCheckBox(x + 7, y + 77, t("gui.refinedstorage:grid_filter.compare_damage"), (compare & IComparer.COMPARE_DAMAGE) == IComparer.COMPARE_DAMAGE);
         compareNBT = addCheckBox(x + 7 + compareDamage.getButtonWidth() + 4, y + 77, t("gui.refinedstorage:grid_filter.compare_nbt"), (compare & IComparer.COMPARE_NBT) == IComparer.COMPARE_NBT);
         compareOredict = addCheckBox(x + 7 + compareDamage.getButtonWidth() + 4 + compareNBT.getButtonWidth() + 4, y + 77, t("gui.refinedstorage:grid_filter.compare_oredict"), (compare & IComparer.COMPARE_OREDICT) == IComparer.COMPARE_OREDICT);
-        toggleMode = addButton(x + 7, y + 71 + 20, 0, 20, "");
+        toggleModFilter = addCheckBox(0, y + 71 + 25, t("gui.refinedstorage:grid_filter.mod_filter"), modFilter);
+        toggleMode = addButton(x + 7, y + 71 + 21, 0, 20, "");
         updateModeButton(mode);
     }
 
     private void updateModeButton(int mode) {
         String text = mode == ItemGridFilter.MODE_WHITELIST ? t("sidebutton.refinedstorage:mode.whitelist") : t("sidebutton.refinedstorage:mode.blacklist");
-        toggleMode.setWidth(fontRendererObj.getStringWidth(text) + 7);
+        toggleMode.setWidth(fontRendererObj.getStringWidth(text) + 12);
         toggleMode.displayString = text;
+        toggleModFilter.xPosition = toggleMode.xPosition + toggleMode.getButtonWidth() + 4;
     }
 
     @Override
@@ -55,7 +60,7 @@ public class GuiGridFilter extends GuiBase {
     @Override
     public void drawForeground(int mouseX, int mouseY) {
         drawString(7, 7, t("gui.refinedstorage:grid_filter"));
-        drawString(7, 114, t("container.inventory"));
+        drawString(7, 115, t("container.inventory"));
     }
 
     @Override
@@ -71,8 +76,10 @@ public class GuiGridFilter extends GuiBase {
         } else if (button == toggleMode) {
             mode = mode == ItemGridFilter.MODE_WHITELIST ? ItemGridFilter.MODE_BLACKLIST : ItemGridFilter.MODE_WHITELIST;
             updateModeButton(mode);
+        } else if (button == toggleModFilter) {
+            modFilter = !modFilter;
         }
 
-        RS.INSTANCE.network.sendToServer(new MessageGridFilterUpdate(compare, mode));
+        RS.INSTANCE.network.sendToServer(new MessageGridFilterUpdate(compare, mode, modFilter));
     }
 }