+++ src/main/java/com/raoulvdberge/refinedstorage/network/MessageGridFilterUpdate.java
@@ -9,25 +9,29 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 public class MessageGridFilterUpdate extends MessageHandlerPlayerToServer<MessageGridFilterUpdate> implements IMessage {
     private int compare;
     private int mode;
+    private boolean modFilter;
 
     public MessageGridFilterUpdate() {
     }
 
-    public MessageGridFilterUpdate(int compare, int mode) {
+    public MessageGridFilterUpdate(int compare, int mode, boolean modFilter) {
         this.compare = compare;
         this.mode = mode;
+        this.modFilter = modFilter;
     }
 
     @Override
     public void fromBytes(ByteBuf buf) {
         compare = buf.readInt();
         mode = buf.readInt();
+        modFilter = buf.readBoolean();
     }
 
     @Override
     public void toBytes(ByteBuf buf) {
         buf.writeInt(compare);
         buf.writeInt(mode);
+        buf.writeBoolean(modFilter);
     }
 
     @Override
@@ -35,6 +39,7 @@ public class MessageGridFilterUpdate extends MessageHandlerPlayerToServer<Messag
         if (player.openContainer instanceof ContainerGridFilter) {
             ItemGridFilter.setCompare(((ContainerGridFilter) player.openContainer).getStack(), message.compare);
             ItemGridFilter.setMode(((ContainerGridFilter) player.openContainer).getStack(), message.mode);
+            ItemGridFilter.setModFilter(((ContainerGridFilter) player.openContainer).getStack(), message.modFilter);
         }
     }
 }